Development

+++ src/Core/Chatter.Application/Services/Rooms/RoomService.cs
@@ -18,6 +18,7 @@ public class RoomService : BaseService, IRoomService
     private readonly IInvitationRepository _invitationRepository;
     private readonly UserManager<ChatterUser> _userManager;
     private readonly IUserRepository _userRepository;
+
     public RoomService(IRoomRepository roomRepository, IInvitationRepository invitationRepository,
         UserManager<ChatterUser> userManager, IUserRepository userRepository)
     {
@@ -29,40 +30,48 @@ public class RoomService : BaseService, IRoomService
 
     public async Task<List<RoomDto>> GetRoomsAsync()
     {
-        return await _roomRepository.Query()
+        var rooms = await _roomRepository.Query(true)
             .Include(x => x.RoomChatterUsers)
-            .ProjectToType<RoomDto>(CreateTypeAdapterConfig(3)).ToListAsync();
+            .ToListAsync();
+
+        return rooms.Adapt<List<RoomDto>>();
     }
 
     public async Task<List<RoomDto>> GetRoomsByUserIdAsync(string userId)
     {
-        var roomIds = await _userRepository.Query()
+        var roomIds = await _userRepository.Query(true)
             .Where(x => x.Id == userId)
             .Include(x => x.RoomChatterUsers)
             .SelectMany(x => x.RoomChatterUsers!).Select(x => x.RoomId).ToListAsync();
-        
-        return await _roomRepository.Query()
+
+        var rooms = await _roomRepository.Query()
             .Include(x => x.RoomChatterUsers)
-            .Where(x => roomIds.Contains(x.Id))
-            .ProjectToType<RoomDto>(CreateTypeAdapterConfig(3)).ToListAsync();
+            .Where(x => roomIds.Contains(x.Id)).ToListAsync();
+
+        return rooms.Adapt<List<RoomDto>>();
     }
 
     public async Task<List<RoomDto>> GetPublicRooms()
     {
-        return await _roomRepository.Query()
+        var rooms = await _roomRepository.Query(true)
             .Where(x => x.IsPublic)
-            .Include(x => x.RoomChatterUsers)
-            .ProjectToType<RoomDto>(CreateTypeAdapterConfig(3)).ToListAsync();
+            .Include(x => x.RoomChatterUsers).ToListAsync();
+
+        return rooms.Adapt<List<RoomDto>>();
     }
 
     public async Task<RoomDto?> GetRoomDetailAsync(int roomId)
     {
-        return await _roomRepository.Query()
+        var room = await _roomRepository.Query(true)
             .Where(x => x.Id == roomId)
             .Include(x => x.RoomPermissions)
             .Include(x => x.RoomChatterUsers)
-            .ThenInclude(x => x.ChatterUser)
-            .ProjectToType<RoomDto>(CreateTypeAdapterConfig(3)).FirstOrDefaultAsync();
+            .ThenInclude(x => x.ChatterUser).FirstOrDefaultAsync();
+
+        if (room is null)
+            throw new FriendlyException("Chat bulunamadı");
+
+        return room.Adapt<RoomDto>();
     }
 
     public async Task<RoomDto> CreateRoomAsync(CreateRoomInput createRoomInput)
@@ -86,7 +95,7 @@ public class RoomService : BaseService, IRoomService
         return room.Adapt<RoomDto>(CreateTypeAdapterConfig(3));
     }
 
-    public async Task BlockUserByRoomAsync(int roomId, ChatterUser blockedUser)
+    public Task BlockUserByRoomAsync(int roomId, ChatterUser blockedUser)
     {
         throw new NotImplementedException();
     }
@@ -97,20 +106,20 @@ public class RoomService : BaseService, IRoomService
             .Include(x => x.RoomChatterUsers)
             .ThenInclude(x => x.ChatterUser)
             .FirstOrDefaultAsync(x => x.Id == joinRoomInput.RoomId);
-        
+
         if (room is null)
-            throw new FriendlyException("Oda bulunamadı");
+            throw new FriendlyException("Chat bulunamadı");
 
         var user = await _userManager.FindByIdAsync(joinRoomInput.UserId);
-        
+
         if (user is null)
             throw new FriendlyException("Kullanıcı bulunamadı");
-        
-        if(room.RoomChatterUsers.Select(x => x.ChatterUserId).Contains(user.Id))
-            throw new FriendlyException("Kullanıcı zaten odada");
+
+        if (room.RoomChatterUsers.Select(x => x.ChatterUserId).Contains(user.Id))
+            throw new FriendlyException("Kullanıcı zaten chatte");
 
         if (room.Capacity == room.RoomChatterUsers.Count)
-            throw new FriendlyException("Oda kapasitesi dolu");
+            throw new FriendlyException("Chat kapasitesi dolu");
 
         var roomChatterUser = new RoomChatterUser()
         {
@@ -118,34 +127,45 @@ public class RoomService : BaseService, IRoomService
             RoomId = room.Id,
         };
         room.RoomChatterUsers?.Add(roomChatterUser);
+
+        var roomPermission = new RoomPermission()
+        {
+            ChatterUserId = user.Id,
+            RoomId = room.Id,
+            PermissionType = ChatPermissionType.Chatter
+        };
+        room.RoomPermissions?.Add(roomPermission);
+
         _roomRepository.Update(room);
     }
 
     public async Task LeaveRoomAsync(LeaveRoomInput leaveRoomInput)
     {
+        // TODO : AsnoTracking yapmayınca Include conditionı çalışmıyor.
         var room = await _roomRepository.Query()
             .Include(x => x.RoomPermissions)
             .Include(x => x.RoomChatterUsers)
             .ThenInclude(x => x.ChatterUser)
             .FirstOrDefaultAsync(x => x.Id == leaveRoomInput.RoomId);
-        
+
         if (room is null)
-            throw new FriendlyException("Oda bulunamadı");
-        
+            throw new FriendlyException("Chat bulunamadı");
+
         var user = await _userManager.FindByIdAsync(leaveRoomInput.UserId);
         if (user is null)
             throw new FriendlyException("Kullanıcı bulunamadı");
-        
-        if(!room.RoomChatterUsers.Select(x => x.ChatterUserId).Contains(user.Id))
-            throw new FriendlyException("Kullanıcı zaten odada değil");
+
+        if (!room.RoomChatterUsers.Select(x => x.ChatterUserId).Contains(user.Id))
+            throw new FriendlyException("Kullanıcı zaten chatte değil");
 
         if (room.RoomPermissions.FirstOrDefault(x => x.ChatterUserId == user.Id)?.PermissionType ==
             ChatPermissionType.Admin &&
             room.RoomPermissions.Count(x => x.PermissionType == ChatPermissionType.Admin) == 1)
-            throw new FriendlyException("Odada başka admin yok. Lütfen başka bir admin atayın.");
-            
+            throw new FriendlyException("Chatte başka admin yok. Lütfen başka bir admin atayın.");
+
         var roomChatterUser = room.RoomChatterUsers.First(x => x.ChatterUserId == user.Id);
-        room.RoomChatterUsers?.Remove(roomChatterUser);
+        roomChatterUser.IsDeleted = true;
+        room.RoomPermissions?.Remove(room.RoomPermissions.First(x => x.ChatterUserId == user.Id));
         _roomRepository.Update(room);
     }
 
@@ -154,18 +174,19 @@ public class RoomService : BaseService, IRoomService
         var room = await _roomRepository.Query()
             .Include(x => x.RoomPermissions)
             .FirstOrDefaultAsync(x => x.Id == deleteRoomInput.RoomId);
-        
+
         if (room is null)
-            throw new FriendlyException("Oda bulunamadı");
-        
+            throw new FriendlyException("Chat bulunamadı");
+
         var user = await _userManager.FindByIdAsync(deleteRoomInput.UserId);
         if (user is null)
             throw new FriendlyException("Kullanıcı bulunamadı");
 
-        if (room.RoomPermissions.FirstOrDefault(x => x.ChatterUserId == deleteRoomInput.UserId)?.PermissionType !=
-            ChatPermissionType.Admin)
-            throw new FriendlyException("Odayı silme yetkiniz yok");
-        
+        if (!(IsFullAdmin(deleteRoomInput.UserId) || room.RoomPermissions
+                    .FirstOrDefault(x => x.ChatterUserId == deleteRoomInput.UserId)?.PermissionType ==
+                ChatPermissionType.Admin))
+            throw new FriendlyException("Chati silme yetkiniz yok");
+
         _roomRepository.Delete(room);
     }
 
@@ -174,20 +195,97 @@ public class RoomService : BaseService, IRoomService
         var room = await _roomRepository.Query()
             .Include(x => x.RoomPermissions)
             .FirstOrDefaultAsync(x => x.Id == editRoomInput.Id);
-        
+
         if (room is null)
-            throw new FriendlyException("Oda bulunamadı");
-        
+            throw new FriendlyException("Chat bulunamadı");
+
         var user = await _userManager.FindByIdAsync(editRoomInput.UserId);
         if (user is null)
             throw new FriendlyException("Kullanıcı bulunamadı");
 
-        if (room.RoomPermissions.FirstOrDefault(x => x.ChatterUserId == editRoomInput.UserId)?.PermissionType !=
-            ChatPermissionType.Admin)
-            throw new FriendlyException("Odayı güncelleme yetkiniz yok");
+        if (!(IsFullAdmin(editRoomInput.UserId) || room.RoomPermissions
+                    .FirstOrDefault(x => x.ChatterUserId == editRoomInput.UserId)?.PermissionType ==
+                ChatPermissionType.Admin))
+            throw new FriendlyException("Chat ayarlarını güncelleme yetkiniz yok");
 
         editRoomInput.Adapt(room);
-        
+
         _roomRepository.Update(room);
+    }
+
+    public async Task AddPermissionToRoomAsync(AddPermissionToRoomInput addPermissionToRoomInput)
+    {
+        var room = await _roomRepository.Query()
+            .Include(x => x.RoomPermissions)
+            .FirstOrDefaultAsync(x => x.Id == addPermissionToRoomInput.RoomId);
+
+        if (room is null)
+            throw new FriendlyException("Chat bulunamadı");
+
+        var requesterUser = await _userManager.FindByIdAsync(addPermissionToRoomInput.RequestedUserId);
+        if (requesterUser is null)
+            throw new FriendlyException("İstek atan kullanıcı bulunamadı");
+
+        if (!(IsFullAdmin(addPermissionToRoomInput.RequestedUserId) || room.RoomPermissions
+                    .FirstOrDefault(x => x.ChatterUserId == addPermissionToRoomInput.RequestedUserId)?.PermissionType ==
+                ChatPermissionType.Admin))
+            throw new FriendlyException("Chate izin verme yetkiniz yok");
+
+        var newChatterUser = await _userManager.FindByIdAsync(addPermissionToRoomInput.ChatterUserId);
+        if (newChatterUser is null)
+            throw new FriendlyException("İşlem yapılmak istenilen kullanıcı bulunamadı");
+
+        var existPermission =
+            room.RoomPermissions.FirstOrDefault(x => x.ChatterUserId == addPermissionToRoomInput.ChatterUserId);
+
+        if (existPermission is null)
+            throw new FriendlyException("İşlem yapılmak istenilen kullanıcı chatte değil.");
+
+        existPermission!.PermissionType = addPermissionToRoomInput.PermissionType;
+
+        _roomRepository.Update(room);
+    }
+
+    public async Task RemoveUserInRoomAsync(RemoveUserInRoomInput removeUserInRoomInput)
+    {
+        var room = await _roomRepository.Query()
+            .Include(x => x.RoomPermissions)
+            .Include(x => x.RoomChatterUsers)
+            .FirstOrDefaultAsync(x => x.Id == removeUserInRoomInput.RoomId);
+
+        if (room is null)
+            throw new FriendlyException("Chat bulunamadı");
+
+        var requesterUser = await _userManager.FindByIdAsync(removeUserInRoomInput.RequestedUserId);
+        if (requesterUser is null)
+            throw new FriendlyException("İstek atan kullanıcı bulunamadı");
+
+        if (!(IsFullAdmin(removeUserInRoomInput.RequestedUserId) || room.RoomPermissions
+                    .FirstOrDefault(x => x.ChatterUserId == removeUserInRoomInput.RequestedUserId)?.PermissionType ==
+                ChatPermissionType.Admin))
+            throw new FriendlyException("Chatten atma yetkiniz yok");
+        
+        var deletedInRoomChatterUser = await _userManager.FindByIdAsync(removeUserInRoomInput.ChatterUserId);
+        if (deletedInRoomChatterUser is null)
+            throw new FriendlyException("İşlem yapılmak istenilen kullanıcı bulunamadı");
+
+        var existPermission =
+            room.RoomPermissions.FirstOrDefault(x => x.ChatterUserId == removeUserInRoomInput.ChatterUserId);
+
+        if (existPermission is null)
+            throw new FriendlyException("İşlem yapılmak istenilen kullanıcı chatte değil.");
+        
+        var roomChatterUser = room.RoomChatterUsers.First(x => x.ChatterUserId == removeUserInRoomInput.ChatterUserId);
+        roomChatterUser.IsDeleted = true;
+        room.RoomPermissions?.Remove(room.RoomPermissions.First(x => x.ChatterUserId == removeUserInRoomInput.ChatterUserId));
+        _roomRepository.Update(room);   
+    }
+
+    private bool IsFullAdmin(string userId)
+    {
+        var user = _userManager.FindByIdAsync(userId).Result;
+        var userRoles = _userManager.GetRolesAsync(user).Result;
+
+        return userRoles.Any(x => x == ChatPermissionType.Admin.ToString());
     }
 }
